quote unquoted special parameters
Previously, special parameters like `?`, `#` or `$` were typically not quoted
because it was assumed that they could never contain any field terminators for
field splitting.

This is however only the case with default values of the shell variable `IFS`.
For example, if `IFS` is set to `2`, a value of `123` in the parameter `?` would
yield the two fields `1` and `3` (in contexts were field splitting is
performed).

While this couldn’t happen in this script (because `IFS` is sanitised in its
beginning), it shall nevertheless be changed to be quoted, which would prevent
any field splitting.

Signed-off-by: Christoph Anton Mitterer <mail@christoph.anton.mitterer.name>